Do type-casting at the very end.

+++ tetrad-gui/src/main/java/edu/cmu/tetradapp/Tetrad.java
@@ -159,11 +159,11 @@ public final class Tetrad implements PropertyChangeListener {
             @Override
             public Dimension getPreferredSize() {
                 GraphicsDevice graphicsDevice = GraphicsEnvironment.getLocalGraphicsEnvironment().getDefaultScreenDevice();
-                int height = (int) (graphicsDevice.getDisplayMode().getHeight() * 0.75);
-                int width = (int) (height * 1.333333333);
+                double height = graphicsDevice.getDisplayMode().getHeight() * 0.75;
+                double width = height * 1.333333333;
 
 //                return Toolkit.getDefaultToolkit().getScreenSize();
-                return new Dimension(width, height);
+                return new Dimension((int) width, (int) height);
             }
 
         };
